# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  domain.Category:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
  domain.CreateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  domain.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  domain.LoginResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/domain.UserResponse'
    type: object
...
  domain.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  domain.RefreshResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
    type: object
...
  domain.UpdateCategoryRequest:
    properties:
      name:
        type: string
    required:
    - name
    type: object
...
  domain.UserResponse:
    properties:
      full_name:
        type: string
      id:
        type: string
      role:
        type: string
      username:
        type: string
    type: object
  response.ErrorDetail:
    properties:
      field:
        type: string
      message:
        type: string
    type: object
  response.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      errors:
        items:
          $ref: '#/definitions/response.ErrorDetail'
        type: array
      message:
        example: Terjadi kesalahan
        type: string
      success:
        example: false
        type: boolean
    type: object
  response.Meta:
    properties:
      has_next:
        type: boolean
      has_prev:
        type: boolean
      limit:
        type: integer
      page:
        type: integer
      total_pages:
        type: integer
      total_records:
        type: integer
    type: object
  response.PaginatedResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        type: string
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        example: true
        type: boolean
    type: object
  response.SuccessResponse:
    properties:
      code:
        example: 200
        type: integer
      data: {}
      message:
        example: Operasi berhasil
        type: string
      success:
        example: true
        type: boolean
    type: object
host: localhost:8080
info:
  contact:
    name: Dev Team Pramuka CAT
    url: https://github.com/odealidj/pramuka-CAT
  description: REST API untuk platform Computer Assisted Test (CAT) Pramuka.
  termsOfService: http://swagger.io/terms/
  title: Pramuka CAT API
  version: "1.0"
paths:
  /admin/categories:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/domain.Category'
                  type: array
              type: object
      security:
      - BearerAuth: []
      summary: Daftar Kategori Soal
      tags:
      - Admin - Kategori
    post:
      consumes:
      - application/json
      parameters:
      - description: Nama Kategori
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.CreateCategoryRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Category'
              type: object
      security:
      - BearerAuth: []
      summary: Buat Kategori Soal
      tags:
      - Admin - Kategori
  /admin/categories/{id}:
    delete:
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
      security:
      - BearerAuth: []
      summary: Hapus Kategori Soal
      tags:
      - Admin - Kategori
    put:
      consumes:
      - application/json
      parameters:
      - description: ID Kategori
        in: path
        name: id
        required: true
        type: integer
      - description: Nama Baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.UpdateCategoryRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.Category'
              type: object
      security:
      - BearerAuth: []
      summary: Update Kategori Soal
      tags:
      - Admin - Kategori
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Autentikasi pengguna dan mendapatkan Access Token & Refresh Token
      parameters:
      - description: Kredensial Login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.LoginResponse'
              type: object
        "401":
          description: Unauthorized
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Login
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Memperbarui Access Token menggunakan Refresh Token yang masih berlaku
      parameters:
      - description: Refresh Token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/domain.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            allOf:
            - $ref: '#/definitions/response.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/domain.RefreshResponse'
              type: object
        "401":
          description: Unauthorized
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Refresh Token
      tags:
      - Auth
  /protected/auth/logout:
    post:
      description: Membatalkan sesi aktif dan menginvalidasi token
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "401":
          description: Unauthorized
          headers:
            X-Trace-Id:
              description: Trace ID untuk distributed tracing. Gunakan nilai ini di
                Jaeger UI (klik Lookup by Trace ID) untuk menemukan detail request
                ini.
              type: string
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token JWT dengan format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"